requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
astropy=6.0.1=pypi_0
astropy-iers-data=0.2024.4.29.0.28.48=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
candle=0.0.1=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
deepchem=2.8.0=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.62.2=pypi_0
h5py=3.11.0=pypi_0
idna=3.7=pypi_0
joblib=1.4.0=pypi_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
patsy=0.5.6=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py310h06a4308_0
protobuf=3.19.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pyerfa=2.0.1.4=pypi_0
pyparsing=3.1.2=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
rdkit=2023.9.6=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
statsmodels=0.14.2=pypi_0
sympy=1.12=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-gpu=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.12=h1ccaba5_0
tqdm=4.66.2=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
werkzeug=3.0.2=pypi_0
wheel=0.41.2=py310h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
